LayoutTests/imported/w3c/web-platform-tests/html/canvas/element/text/2d.text.drawing.style.wordSpacing.change.font-expected.txt

2d.text.drawing.style.wordSpacing.change.font
Set word spacing and word spacing to font dependent value and verify it works after font change.
Actual output:

FAIL Set word spacing and word spacing to font dependent value and verify it works after font change. assert_equals: ctx.letterSpacing === '0px' (got [undefined], expected 0px[string]) expected (string) "0px" but got (undefined) undefined
